Revert "Added variable *zap-input-newlines* which, if set, tells the command parser to"
This reverts commit 8c149a11aa20085166db906fb90d534316f6c194.

Realized that this is a dumb feature.  I mainly wanted it so I could write
comments at the end of lines that will continue of the next line.  That just
isn't possible with shell scripts, so I gave up.  You have to write your
embedded shell scripts like you would a shell script/