There is an issue with at least CLISP.  The CLISP reader slurps one whitespace character after reading an atom.  This means that that whitespace isn't passed to the shell and causes problems.  Other readers don't seem to do this.
Changed the READs to READ-PRESERVING-WHITESPACE in the reader macro.  This
really doesn't do anything, because as the standard states, if reading in a
recursive environment, the two are identical.  However, the standard also seems
a little confused on the matter, see the example in the hyperspec.  For now I
have a special hack for CLISP where those reads are performed with :recursive-p
passed as NIL.  I'm not sure what effect this will have, but for now it seems to
work.