New reader syntax for controlling how the command in executed.
You can specify letters after the '>' but before the form that control how the
command is executed.  'p' means predicate mode (currently unimplemented), 'e'
and 'w' specify to raise errors or warnings on output on stderr (currently
unimplemented), '&' tells the command to run in the background (makes all of the
other options irrelevant), and x (with an optional integer after it) specifies
the exit code that signals no error (precede with a '!' to indicate that any
exit code other than that integer signals no error).