Better process control (attempted).
A try at allowing the user to interrupt the shell process via C-c C-c from Slime
or C-c from shell, and to have it kill the process if you leave the block but
have specified that you wish to wait (i.e. via non-local exit).

Neither seem to work from Slime, but at least the interrupt seems to have
promise from the shell.  This leads me to believe that the Slime error handling
is getting in the way.  Still not sure what to do about it.