Cleaning up the interface.
There are two main function interfaces, CMD and CMD-BG.

CMD runs the command in the foreground and return either a stream or a streang
depending on the keyword parameter you pass via :output.

CMD-BG runs the command in the background and returns a cmd-process structure.

A third function is exported called CMD-P.  This runs a CMD in the foreground,
like CMD, but does so with a setup that makes the shell command work like a
pseudo-predicate.