A very deep reworking.
The way this works has changed in many, sometimes subtle, ways.  Sorry.  Support
for all but SBCL has been officially temporarily dropped.  Any non-threaded Lisp
imp (e.g. CLISP, but also certain builds of other imps) will only be able to do
a subset of this, though I plan to come up with an emulation using Gray Streams.

Fair handling of output on standard output and standard error.  First complete
line is printed first.

%CMD is now much more bare-bones.  Though it is still a bit up in the air, it is
shaping up that it need only accept nil or :stream for :output and :error.

CMD now always returns three values, the output as a string, the error as a
string, and the exit code.

CMD-BG now always returns two values, a process structure that contains the
output streams and a thread that is handling the background process.  The
streams need not be closed as they are string streams.

CMD-P should work the same way, roughly.

You can now have Lisp raise an error whenever there is output to stderr (or it
warn) using :on-error-output.  This means that errors will be raised near the
time they occur rather than at the end of the computation.  The computation
continues even when the debugger is invoked; I'm working on making this work
more like normal Lisp (pauses with the option of resume).

Error messages have error output as part of their error message.  This is done
even if you have redirected standard error to output.

Streams are cleaned up and shell processes are kill upon leaving the monitoring
thread.  This means that there will be no more run-away sub-processes that need
to be killed from the shell.

Jobs are logged and can be retrieved via the ALL-JOBS function.  Jobs are
removed as they complete.

The reader macro now takes more values on the control string.  Namely it allows
for splitting and 'e' and 'w', which if present, turn on the :on-error-output
:error and :on-error-output :warn options, respectively.