Added some support for ECL back into the library.
ECL support for output on stderr seems to be broken.  So we allow only nil and
:output until that is resolved.  This means, also, that on-error-output options
will not work.  We can actually work around this without having tinkering with
ECL internals, but I would rather see the change there.  I need a bug report or
patch.

The ECL conditional reader macro isn't working the way I would expect.  It seems
that it might disregard more than a single form after a failed condition, or
something.  In the mean time I reorganized the conditionals so that ECL can
hobble along.