Bug fixes: close inlet of pipe when all output has been read.
This is necessary for when you are reading from a pipe as the pipe outlet
doesn't show EOF until the inlet is closed.  So you have to close it if you
don't want it to hang on read.

Related, we switched from make-two-way-pipe to make-pipe in order to simplify
the closing of the input pipe.